backend-worker/requirements.txt
flask==3.0.0
google-api-python-client==2.111.0
google-auth==2.25.2
requests==2.31.0
pytz==2024.1
gunicorn==21.2.0
